Initial clean-up for IDE listening for .editorconfig changes
