+++ src/EditorFeatures/Core.Wpf/EditorFeatures.Wpf.csproj
@@ -9,7 +9,6 @@
     <AssemblyName>Microsoft.CodeAnalysis.EditorFeatures.Wpf</AssemblyName>
     <TargetFramework>net46</TargetFramework>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <DefineConstants>$(DefineConstants);EDITORFEATURESWPF</DefineConstants>
   </PropertyGroup>
   <ItemGroup Label="Project References">
     <ProjectReference Include="..\..\Compilers\Core\Portable\CodeAnalysis.csproj" />
